# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A small in-memory to-do list API built for FlyRank Backend Track W2 A1. No database — data resets on restart."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development server" }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API description",
        "description": "Returns basic info about this API.",
        "responses": {
          "200": { "description": "API info" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Used to confirm the server is alive.",
        "responses": {
          "200": { "description": "Server is healthy" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "description": "Returns every task. Optionally filter with ?done=true|false or ?search=word.",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean" },
            "description": "Filter tasks by completion status"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter tasks whose title contains this text"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tasks",
            "content": {
              "application/json": {
                "example": [{ "id": 1, "title": "Buy milk", "done": false }]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "description": "Creates a new task from the JSON body. title is required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Missing or invalid title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "The task" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Replaces title and/or done. At least one field is required.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated task" },
          "400": { "description": "Invalid body" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted, no content" },
          "404": { "description": "Task not found" }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "(extra) Task statistics",
        "responses": {
          "200": { "description": "Total, done, and open counts" }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "(extra) Reset to seed data",
        "responses": {
          "200": { "description": "Tasks reset" }
        }
      }
    }
  }
}